python/requirements.txt
# Vulnerable dependencies for CVE testing (Python 3.11 compatible)
fastapi==0.68.0
uvicorn==0.15.0
jinja2==2.11.0
python-dotenv>=1.0.0
pydantic==1.8.2
pytest>=8.3.4
pytest-cov>=4.1.0
httpx>=0.23.0
requests==2.25.1
pillow==8.3.2
pyyaml==6.0
cryptography==3.4.8
django==3.2.15
flask==2.0.3
werkzeug==2.0.3
click==7.1.2
urllib3==1.26.5
setuptools==65.5.0
wheel==0.37.1
